requirements.txt
asgiref==3.8.1
click==8.1.7
colorama==0.4.6
dj-database-url==2.1.0
Django==4.2.11
djangorestframework==3.15.1
joblib==1.3.2
nltk==3.8.1
numpy==1.26.4
psycopg2==2.9.9
regex==2023.12.25
sqlparse==0.4.4
tqdm==4.66.2
typing-extensions==4.10.0
tzdata==2024.1
